requirements.txt
pandas
matplotlib
numpy
torch
torchvision
scipy
scikit-learn
seaborn
openpyxl
pillow
